Separation of concern baby
